SVO: Added synchronization between the 2 threads
- By default, the performance of the SVO is given by the processing
time of the tracker (main) thread. In the slambench implementation of
SVO, the main thread finishes processing each frame so fast that it does
not give enough time to the mapper thread to finish its task. A wait-notify
mechanism has been implemented so that the tracker thread waits for the
mapper to finish processing before proceeding to the next frame.